Disallow cross-site requests by default
The default behavior for *CORS* was to allow all cross-site requests,
because the default value was set to `["*"]`.
This is obviously a very bad default from security point of view.

The default behavior should be to disallow all cross-site requests.
For an application like MXCuBE-Web,
allowing same-site requests only is usually good enough.