adds support for 'create_point' task flag
Adds support for 'create_point' flag for custom tasks. If this
flag is set, create a 2D-point before showing the task dialog.

This make it possible to implement same behaviour for custom tasks as
for the built-in 'Data Collection (Limited OSC)' task. I.e. a 2D
point is created where user clicked, and the collection is performed
at that point.

squash formatting